+++ src/lib.rs
@@ -1,4 +1,5 @@
 #![no_std]
+#![deny(missing_docs)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
 //! # io-webdav